Base config + home-manager
This is just my base config with the addition of my home.nix file (home-manager configuration). I will be managing this more and more as time goes on. 

Next goal is theming. Once theming is configured properly, I will move on to getting a Data Science/ML setup in the works.